Move speculators weight map and remap functions to vllm/transformers_utils/configs/speculators_eagle.py
